fast_api/requirements.txt
fastapi==0.110.1
pycryptodome==3.20.0
uvicorn==0.29.0